requirements.txt
streamlit
PyMuPDF
Pillow
pytesseract
sentence-transformers
faiss-cpu
numpy
langchain
groq
python-dotenv